Adding Int64 enum to test support

+++ db/migrations/20200324211956_create_issue.cr
@@ -2,7 +2,7 @@ class CreateIssue::V20200324211956 < Avram::Migrator::Migration::V1
   def migrate
     create table_for(Issue) do
       primary_key id : Int64
-      add status : Int32
+      add status : Int64
       add role : Int32
       add_timestamps
     end
